backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
python-multipart==0.0.6
firebase-admin==6.2.0
openmeteo-requests==1.3.0
requests-cache==1.1.1
retry-requests==2.0.0
